[UPD] nested_model generation enhancements
